pyproject.toml
[tool.poetry]
name = "google-api"
version = "1.01"
description = "Package to interact with various Google APIs"
authors = ["Andre <andre.kamarudin@fairprice.com.sg>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
db-dtypes = ">=1.3.0,<2.0.0"
tabulate = ">=0.9.0,<1.0.0"
google-api-core = ">=2.0.0,<3.0.0"
google-cloud-bigquery = ">=3.0.0,<4.0.0"
google-auth = ">=2.0.0,<3.0.0"
tqdm = ">=4.0.0,<5.0.0"
google-api-python-client = ">=2.0.0,<3.0.0"
pydantic = ">=2.0.0,<3.0.0"
loguru = ">=0.0.0,<1.0.0"
colorama = ">=0.0.0,<1.0.0"
icecream = ">=2.0.0,<3.0.0"
pandas = ">=2.0.0,<3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"